Moved HSV color to types/color.ts, removed duplicated events
