+++ packages/react-components/react-color-picker-preview/library/src/components/ColorArea/useColorArea.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useId, slot, useMergedRefs, mergeCallbacks, getIntrinsicElementProps } from '@fluentui/react-utilities';
 import type { ColorAreaProps, ColorAreaState } from './ColorArea.types';
-import type { HsvColor } from '../ColorPicker/ColorPicker.types';
+import type { HsvColor } from '../../types/color';
 import { colorAreaCSSVars } from './useColorAreaStyles.styles';
 import { useEventCallback, useControllableState } from '@fluentui/react-utilities';
 import { useFluent_unstable as useFluent } from '@fluentui/react-shared-contexts';
