+++ packages/react-components/react-color-picker-preview/library/src/types/color.ts
@@ -1,0 +1,6 @@
+export type HsvColor = {
+  h: number;
+  s: number;
+  v: number;
+  a?: number;
+};
